@@ -3,15 +3,12 @@ package com.samuelberrien.phyvr;
 import android.app.NativeActivity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
-import android.widget.CheckBox;
 import android.widget.NumberPicker;
 
 import com.samuelberrien.phyvr.controls.ControlActivity;
 import com.samuelberrien.phyvr.controls.GamePadActivity;
-import com.samuelberrien.phyvr.game.PlayActivity;
-import com.samuelberrien.phyvr.game.MyGvrActivity;
 
 public class MainActivity extends AppCompatActivity implements NumberPicker.OnValueChangeListener {
 
